Refactor: Changed scrollToTop reference
